refactor(pkg/civitai): narrow public SDK surface to read/download only
Promoting the whole former internal/api package to pkg/civitai (#172) made
every exported symbol a permanent compat contract — including the CLI-internal
App Blocks machinery (Submitter/Verifier/Withdrawer, dev-tunnel + dev-token,
OAuth device flow, ForgejoCloneInfo, the route/path consts, ClientID, and the
token-scope bit flags), which no SDK consumer needs.

Move the App-Blocks-only surface into a new CLI-internal package
internal/appapi (its own Client with the auth/submit plumbing), leaving
pkg/civitai's exported API to exactly the read/download contract: Client,
Reader, Downloader, the typed model/version/image/tag/creator/user/article/
collection structs, PrimaryFile, the ErrX error kinds + Tag/TagStatus,
TokenSource/StaticToken, and New/NewWithSource.

- internal/appapi reuses only pkg/civitai's EXPORTED helpers (TokenSource,
  StaticToken, Tag/TagStatus + error kinds); it duplicates the small
  body-read cap helper so pkg/civitai needs ZERO new exports. pkg/civitai
  imports nothing under internal/.
- civitai.New/NewWithSource drop the App-Blocks submitPath parameter
  (New(baseURL, token) / NewWithSource(baseURL, src)); the submit path now
  lives on appapi.Client.
- internal/cmd/*, internal/auth, and cmd/civitai updated to use appapi for
  the App-Blocks operations; read/download commands keep using pkg/civitai.
- App-Blocks tests moved alongside the code into internal/appapi; the
  read/download tests and the external-consumer example_test.go guard stay in
  pkg/civitai (adapted to the 2-arg constructors, unchanged assertions).

No behavior change to either surface. build/vet/test/gofmt/golangci-lint all
clean.
