+++ internal/appapi/dev_token_test.go
@@ -1,4 +1,4 @@
-package civitai
+package appapi
 
 import (
 	"context"
