+++ internal/appapi/oauth_msg_test.go
@@ -1,4 +1,4 @@
-package civitai
+package appapi
 
 import (
 	"strings"
