+++ internal/auth/source_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/civitai/cli/internal/appapi"
 	"github.com/civitai/cli/internal/config"
 	"github.com/civitai/cli/pkg/civitai"
 )
@@ -14,11 +15,11 @@ import (
 type stubRefresher struct {
 	calls int
 	sawRT string
-	resp  *civitai.TokenResponse
+	resp  *appapi.TokenResponse
 	rerr  error
 }
 
-func (s *stubRefresher) Refresh(_ context.Context, rt string) (*civitai.TokenResponse, error) {
+func (s *stubRefresher) Refresh(_ context.Context, rt string) (*appapi.TokenResponse, error) {
 	s.calls++
 	s.sawRT = rt
 	if s.rerr != nil {
@@ -47,7 +48,7 @@ func TestSourceExpiredAccessTokenTriggersRefresh(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	ref := &stubRefresher{resp: &civitai.TokenResponse{
+	ref := &stubRefresher{resp: &appapi.TokenResponse{
 		AccessToken: "new-at", ExpiresIn: 3600, RefreshToken: "rotated-rt", Scope: "s",
 	}}
 	src := newWithRefresher(cfg, ref)
@@ -81,7 +82,7 @@ func TestSourceValidTokenSkipsRefresh(t *testing.T) {
 	if err := cfg.SetOAuthTokens("good-at", "rt", time.Now().Add(time.Hour), "s"); err != nil {
 		t.Fatal(err)
 	}
-	ref := &stubRefresher{resp: &civitai.TokenResponse{AccessToken: "nope"}}
+	ref := &stubRefresher{resp: &appapi.TokenResponse{AccessToken: "nope"}}
 	src := newWithRefresher(cfg, ref)
 
 	tok, err := src.Token(context.Background())
@@ -106,7 +107,7 @@ func TestSourceRefreshKeepsOldRefreshTokenWhenNotRotated(t *testing.T) {
 		t.Fatal(err)
 	}
 	// Server returns no refresh_token (no rotation).
-	ref := &stubRefresher{resp: &civitai.TokenResponse{AccessToken: "new-at", ExpiresIn: 3600}}
+	ref := &stubRefresher{resp: &appapi.TokenResponse{AccessToken: "new-at", ExpiresIn: 3600}}
 	src := newWithRefresher(cfg, ref)
 
 	if _, err := src.Token(context.Background()); err != nil {
@@ -129,7 +130,7 @@ func TestSourceRefreshRejectsNonPositiveExpiresIn(t *testing.T) {
 	}
 	// A refresh that returns expires_in<=0 must be treated as an error, not
 	// persisted with an ~now expiry (which would refresh on every call).
-	ref := &stubRefresher{resp: &civitai.TokenResponse{AccessToken: "new-at", ExpiresIn: 0, RefreshToken: "new-rt"}}
+	ref := &stubRefresher{resp: &appapi.TokenResponse{AccessToken: "new-at", ExpiresIn: 0, RefreshToken: "new-rt"}}
 	src := newWithRefresher(cfg, ref)
 
 	if _, err := src.Token(context.Background()); err == nil {
@@ -153,7 +154,7 @@ func TestSourceRefreshPersistsArrayScope(t *testing.T) {
 	}
 	// Scope arriving as the array-normalized string (what the refresh route
 	// yields via the custom Scope unmarshaler) must round-trip into config.
-	ref := &stubRefresher{resp: &civitai.TokenResponse{
+	ref := &stubRefresher{resp: &appapi.TokenResponse{
 		AccessToken: "new-at", ExpiresIn: 3600, RefreshToken: "new-rt", Scope: "33554433",
 	}}
 	src := newWithRefresher(cfg, ref)
@@ -176,7 +177,7 @@ func TestSourcePersonalKeyNeverRefreshes(t *testing.T) {
 	if err := cfg.SetToken("personal-key"); err != nil {
 		t.Fatal(err)
 	}
-	ref := &stubRefresher{resp: &civitai.TokenResponse{AccessToken: "should-not-be-used"}}
+	ref := &stubRefresher{resp: &appapi.TokenResponse{AccessToken: "should-not-be-used"}}
 	src := newWithRefresher(cfg, ref)
 
 	tok, err := src.Token(context.Background())
