+++ internal/cmd/app_withdraw.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/civitai/cli/internal/appapi"
 	"github.com/civitai/cli/internal/auth"
 	"github.com/civitai/cli/internal/config"
 	"github.com/civitai/cli/internal/ui"
@@ -57,7 +58,7 @@ Pass the publish-request id as a positional argument or via --id (find it with
 				return fmt.Errorf("a publish-request id is required — pass it as an argument or with --id (find it via `civitai app status`)")
 			}
 
-			client := civitai.NewWithSource(cfg.BaseURL(), auth.New(cfg), "")
+			client := appapi.NewWithSource(cfg.BaseURL(), auth.New(cfg), "")
 			ctx := context.Background()
 
 			if err := client.WithdrawRequest(ctx, id); err != nil {
