+++ internal/cmd/download_cancel_test.go
@@ -35,7 +35,7 @@ func TestDownloadOneCancelCleansPart(t *testing.T) {
 
 	dir := t.TempDir()
 	target := filepath.Join(dir, "m.safetensors")
-	dl := civitai.New(srv.URL, "", "")
+	dl := civitai.New(srv.URL, "")
 	dl.AllowPrivateDownloadHosts = true // httptest binds plain-http loopback
 	f := civitai.ModelVersionFile{Name: "m.safetensors", Type: "Model", DownloadURL: srv.URL + "/dl"}
 	o := &downloadOpts{noVerify: true}
